UI Update
To simplify scope for this first "release", I changed the starting storyboard for the project.  Rather than starting on the main tab bar controller, the project now starts right on the Pet LIstings storyboard.  This matches the scope of the Android project.

I also made some updates to the UI on the Pet Listings view controller to more closely match the Android app's UI.  The Detail view controller has not been updated yet, but I anticipate this being updated as well.  I decided to hold off on those changes until the API updates have been made, as there may be some changes to data models, etc.