Re-did the new Pet Details UI
I didn't like how the original design was looking, so I went with something closer to what the original prototype was laying out.  The cells are still static (which could probably be changed at some point), but this makes a little more sense to me.  I dunno, tell me if I'm wrong.